Revert "[3.13] gh-136759: rename `lock.h` to `pylock.h` (GH-137041) (#137075)"
This reverts commit 74503acba7d6c563aeef307ccf2d0cee141838b0.
