--- Misc/NEWS.d/next/C API/2025-07-23-22-30-23.gh-issue-136759.ffB4wO.rst
@@ -1,1 +1,0 @@
-Rename ``lock.h`` to ``pylock.h`` to avoid potential include conflicts.
